move public folder
* issue: there existed a "public" folder in root directory, as you created a new "public" folder in "pages", which is confusing

* move "publick" folder into the root directory
