+++ src/pages/_app.js
@@ -1,7 +1,7 @@
 import GlobalStyles from "../components/GlobalStyles";
 import Layout from "../components/Layout";
 import { Hydrate } from "react-query";
-import "../public/static/fonts/style.css";
+import "../../public/static/fonts/style.css";
 
 export default function MyApp({ Component, pageProps }) {
   return (
